fix types parameters

+++ packages/plasm-utils/src/types.ts
@@ -1,7 +1,7 @@
 /* eslint new-cap: ["error", { "newIsCap": false }] */
 // @ts-check
 // Import the API
-import { Struct, u128, bool } from '@polkadot/types';
+import { Struct, u128, bool, u32 } from '@polkadot/types';
 import { registry } from '@polkadot/react-api';
 
 export class Parameters extends Struct {
@@ -9,7 +9,7 @@ export class Parameters extends Struct {
     super(registry, {
       canBeNominated: 'bool',
       optionExpired: 'u128',
-      optionP: 'u128'
+      optionP: 'u32'
     }, value);
   }
 
@@ -17,7 +17,7 @@ export class Parameters extends Struct {
     return new Parameters({
       canBeNominated: new bool(registry, true),
       optionExpired: new u128(registry, 0),
-      optionP: new u128(registry, 0)
+      optionP: new u32(registry, 0)
     });
   }
 
@@ -30,6 +30,6 @@ export const types = {
   Parameters: {
     canBeNominated: 'bool',
     optionExpired: 'u128',
-    optionP: 'u128'
+    optionP: 'u32'
   }
 };
